+++ Classes/Domain/Model/BaseModel.php
@@ -235,7 +235,6 @@ class BaseModel extends DataModel implements DomainInterface, DynamicTableInterf
             $field = tx_rnbase_util_TCA::getTstampFieldForTable($tableName);
             if (!$this->isPropertyEmpty($field)) {
                 $tstamp = (int) $this->getProperty($field);
-                tx_rnbase::load('tx_rnbase_util_Dates');
                 $datetime = tx_rnbase_util_Dates::getDateTime(
                     '@'.$tstamp,
                     $timezone