+++ Classes/Domain/Model/MediaModel.php
@@ -1,8 +1,13 @@
 <?php
+
+namespace Sys25\RnBase\Domain\Model;
+
+use Sys25\RnBase\Utility\Misc;
+
 /***************************************************************
  *  Copyright notice
  *
- *  (c) 2008-2016 Rene Nitzsche (rene@system25.de)
+ *  (c) 2008-2021 Rene Nitzsche (rene@system25.de)
  *  All rights reserved
  *
  *  This script is part of the TYPO3 project. The TYPO3 project is
@@ -22,15 +27,8 @@
  *  This copyright notice MUST APPEAR in all copies of the script!
  ***************************************************************/
 
-// Die Datenbank-Klasse
-tx_rnbase::load('tx_rnbase_model_base');
-
-class tx_rnbase_model_media extends tx_rnbase_model_base
+class MediaModel extends BaseModel
 {
-    public $uid;
-
-    public $record;
-
     public function __construct($rowOrUid)
     {
         if (is_object($rowOrUid)) {
@@ -45,7 +43,7 @@ class tx_rnbase_model_media extends tx_rnbase_model_base
     private function initMedia($media)
     {
         // Bei FAL steckt in Media eine Referenz
-        if ($media instanceof TYPO3\CMS\Core\Resource\FileReference) {
+        if ($media instanceof \TYPO3\CMS\Core\Resource\FileReference) {
             $this->initFalReference($media);
         } else {
             $this->initFalFile($media);
@@ -53,37 +51,37 @@ class tx_rnbase_model_media extends tx_rnbase_model_base
     }
 
     /**
-     * @param TYPO3\CMS\Core\Resource\File $media
+     * @param \TYPO3\CMS\Core\Resource\File $media
      */
     private function initFalFile($media)
     {
-        $this->record = $media->getProperties();
-        $this->uid = $media->getUid();
-        $this->record['fal_file'] = '1'; // Das wird per TS ausgewertet. Die UID ist KEINE Referenz
-        $this->record['uid_local'] = $media->getUid();
-        $this->record['file_path'] = $media->getPublicUrl();
-        $this->record['file_abs_url'] = tx_rnbase_util_Misc::getIndpEnv('TYPO3_SITE_URL').$this->record['file_path'];
+        $this->setProperty($media->getProperties());
+        $this->setUid($media->getUid());
+        $this->setProperty('fal_file', '1'); // Das wird per TS ausgewertet. Die UID ist KEINE Referenz
+        $this->setProperty('uid_local', $media->getUid());
+        $this->setProperty('file_path', $media->getPublicUrl());
+        $this->setProperty('file_abs_url', Misc::getIndpEnv('TYPO3_SITE_URL').$this->getProperty('file_path'));
     }
 
     /**
-     * @param TYPO3\CMS\Core\Resource\FileReference $media
+     * @param \TYPO3\CMS\Core\Resource\FileReference $media
      */
     private function initFalReference($media)
     {
-        $this->record = $media->getProperties();
+        $this->setProperty($media->getProperties());
         // Wir verwenden hier die UID der Referenz
-        $this->uid = $media->getUid();
-        $this->record['uid'] = $media->getUid();
-        $this->record['file_path'] = $media->getPublicUrl();
-        $this->record['file_abs_url'] = tx_rnbase_util_Misc::getIndpEnv('TYPO3_SITE_URL').$this->record['file_path'];
+        $this->setUid($media->getUid());
+        $this->setProperty('uid', $media->getUid());
+        $this->setProperty('file_path', $media->getPublicUrl());
+        $this->setProperty('file_abs_url', Misc::getIndpEnv('TYPO3_SITE_URL').$this->getProperty('file_path'));
     }
 
     private function initAdditionalData()
     {
-        $this->record['file'] = urldecode($this->record['file_path'].$this->record['file_name']);
+        $this->setProperty('file', urldecode($this->getProperty('file_path').$this->getProperty('file_name')));
         // Some more file fields are useful
-        $this->record['file1'] = $this->record['file'];
-        $this->record['thumbnail'] = $this->record['file'];
+        $this->setProperty('file1', $this->getProperty('file'));
+        $this->setProperty('thumbnail', $this->getProperty('file'));
     }
 
     /**
@@ -93,6 +91,6 @@ class tx_rnbase_model_media extends tx_rnbase_model_base
      */
     public function getTableName()
     {
-        return 'tx_dam';
+        return 'sys_file';
     }
 }